### 2025-01-16 #7671 @editorialbot
[PRE REVIEW]: The Causal Testing Framework
**Submitting author:** @jmafoster1 (<a href="http://orcid.org/0000-0001-8233-9873">Michael Foster</a>)
**Repository:** https://github.com/CITCOM-project/CausalTestingFramework
**Branch with paper.md** (empty if default branch):
**Version:** v9.0.2
