### 2025-01-16 #7671 @danielskatz
👋  @VincentStimper & @tedinburgh, would one or both of you be willing to review this submission for JOSS?
