### 2025-01-20 #7671 @danielskatz
👋 @FATelarico & @alberto743 - Would one or both of you be willing to review this JOSS submission?
