### 2025-01-20 #7671 @alberto743
@danielskatz Thank you for your invitation. Unfortunately, I am reviewing other two papers for JOSS. I prefer to first finish the others before getting another one. Please ping me in the forthcoming weeks if you do not find other reviewers.
