### 2025-01-20 #7671 @FATelarico
> 👋 [@FATelarico](https://github.com/FATelarico) & [@alberto743](https://github.com/alberto743) - Would one or both of you be willing to review this JOSS submission?

 👋 I have another review with JOSS open, officially. But I just completed my checklist today. So, I am available
