+++ paper/paper.bib
@@ -83,7 +83,7 @@
   month = {09},
   pagecount = {464},
   publisher = {Cambridge university press},
-  subtitle = {Models, Reasoning, and Infernce},
+  subtitle = {Models, Reasoning, and Inference},
   title = {Causality},
   year = {2009}
 }
