### 2025-01-26 #7671 @VincentStimper
Unfortunately, I cannot review this paper.
