### 2025-01-30 #7671 @danielskatz
I'll go ahead and start this review now
