### 2025-01-30 #7739 @editorialbot
[REVIEW]: The Causal Testing Framework
**Submitting author:** @jmafoster1 (<a href="http://orcid.org/0000-0001-8233-9873">Michael Foster</a>)
**Repository:** https://github.com/CITCOM-project/CausalTestingFramework
**Branch with paper.md** (empty if default branch):
**Version:** v10.0.1
**Archive:** 10.15131/shef.data.28539398.v1
