### 2025-01-30 9ea0511
Can now execute EA tests
