+++ .github/workflows/ci-tests.yaml
@@ -40,3 +40,4 @@ jobs:
       with:
         fail_ci_if_error: true
         token: ${{ secrets.CODECOV_TOKEN }}
+        version: "v0.1.15"
