### 2025-02-07 #7739 @jmafoster1
Thanks @FATelarico for your review. In terms of authorship, I included everyone who has made a substantial contribution to the project, whether they have contributed code or not. I summarise contributions below:
  - Michael Foster: Me, researcher on the CITCoM project and current lead developer
  - Andrew Clark: PhD student who laid the groundwork for this project and initially led development
  - Christopher Wild: RSE, helped with documentation, refactoring, and other non-core research related development
  - Farhad Allian: RSE, helped with documentation, refactoring, and other non-core research related development. Took over when Robert Turner left
  - Robert Turner: RSE, helped with documentation, refactoring, and other non-core research related development. Unfortunately left us fairly early on in the project, but was extremely helpful in directing the initial setup.
  - Richard Somers: PhD student on the project, who led the development of the surrogate modelling modules.
  - Nicholas Latimer: Co-I on CITCoM. He was extremely helpful in providing technical advice for implementing temporal estimation. While he has not directly contributed code to this project, he graciously provided me with Stata and R implementations of the estimation algorithms from his own research for me to generalise and translate into Python, which was also tremendously helpful for testing.
  - Neil Walkinshaw: PI of the CITCoM project. He has not committed code himself, but has directed the research project and advised on both technical and non-technical aspects of the code and its evaluation.
  - Rob Hierons: Co-I on CITCoM. He has not committed code himself, but has directed the research project and advised on both technical and non-technical aspects of the code and its evaluation.

As for your installation issue on Debian, I've never used Debian so I'm afraid I'm not sure. We initially had some issues with pygraphviz requiring  `libgraphviz-dev` on ubuntu, and graphviz to be installed and on the path on Windows. Perhaps it could be something similar, or requiring some other lower level dependency like GCC or G++ or something? Or it could be the 32 bit nature of the machine as I'm not sure we've tried it on that. With no error message, it's very hard to look into it. Is there really no output at all from the failed install? @f-allian, do you have any ideas on this? Is it something that's worth looking into do you think, or should we simply put a note to the effect of "There have been reports on this failing to install on 32 bit Debian. Feel free to try and let us know if you have any issues." in the install notes?
