### 2025-02-07 #7739 @FATelarico
As I saidz there's no defined policy on authorship. So, as long as everyone on the submitting side is comfortable with the authors listed, and given that there's only so much to be inferred from GitHub commits, I ticked also that box.

> Is there really no output at all from the failed install? @f-allian, do you have any ideas on this?

The terminal just crushes, it's possible that there's a message being printed shortly before that, but I wasn't able to make sure of it.

> Is it something that's worth looking into do you think, or should we simply put a note to the effect of "There have been reports on this failing to install on 32 bit Debian. Feel free to try and let us know if you have any issues." in the install notes?

I think that except for some older machines in developing-countries institutions, most user will have a 64-bit machine. So, I am not suggesting much deep analysis of it — especially if no apparent reason comes to mind. I'd suggest adding a note about 32-bit systems for the aforementioned user group, but there's not much more to be done about it.
