### 2025-02-17 b8ace8e
Removed data collector from causal test suite
