### 2025-02-17 2ceb654
Removed datacollector from testing
