+++ examples/covasim_/doubling_beta/example_beta.py
@@ -6,7 +6,6 @@ from causal_testing.specification.causal_dag import CausalDAG
 from causal_testing.specification.scenario import Scenario
 from causal_testing.specification.variable import Input, Output
 from causal_testing.specification.causal_specification import CausalSpecification
-from causal_testing.data_collection.data_collector import ObservationalDataCollector
 from causal_testing.testing.causal_test_case import CausalTestCase
 from causal_testing.testing.causal_test_outcome import Positive
 from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
@@ -52,7 +51,26 @@ def doubling_beta_CATE_on_csv(
 
     # Read in the observational data, perform identification
     past_execution_df = pd.read_csv(observational_data_path)
-    data_collector, _, causal_test_case, causal_specification = setup(past_execution_df)
+
+    # 2. Create variables
+    pop_size = Input("pop_size", int)
+    pop_infected = Input("pop_infected", int)
+    n_days = Input("n_days", int)
+    cum_infections = Output("cum_infections", int)
+    cum_deaths = Output("cum_deaths", int)
+    location = Input("location", str)
+    variants = Input("variants", str)
+    avg_age = Input("avg_age", float)
+    beta = Input("beta", float)
+    contacts = Input("contacts", float)
+
+    # 5. Create a base test case
+    base_test_case = BaseTestCase(treatment_variable=beta, outcome_variable=cum_infections)
+
+    # 6. Create a causal test case
+    causal_test_case = CausalTestCase(
+        base_test_case=base_test_case, expected_causal_effect=Positive, control_value=0.016, treatment_value=0.032
+    )
 
     linear_regression_estimator = LinearRegressionEstimator(
         "beta",
@@ -98,15 +116,6 @@ def doubling_beta_CATE_on_csv(
     # Repeat causal inference after deleting all rows with treatment value to obtain counterfactual inferences
     if simulate_counterfactuals:
         counterfactual_past_execution_df = past_execution_df[past_execution_df["beta"] != 0.032]
-        counterfactual_linear_regression_estimator = LinearRegressionEstimator(
-            "beta",
-            0.032,
-            0.016,
-            {"avg_age", "contacts"},
-            "cum_infections",
-            df=counterfactual_past_execution_df,
-            formula="cum_infections ~ beta + I(beta ** 2) + avg_age + contacts",
-        )
         counterfactual_causal_test_result = causal_test_case.execute_test(estimator=linear_regression_estimator)
 
         results_dict["counterfactual"] = {
@@ -213,59 +222,6 @@ def doubling_beta_CATEs(observational_data_path: str, simulate_counterfactual: b
     all_fig.savefig(outpath_base_str + "all_executions.pdf", format="pdf")
     age_fig.savefig(outpath_base_str + "age_executions.pdf", format="pdf")
     age_contact_fig.savefig(outpath_base_str + "age_contact_executions.pdf", format="pdf")
-
-
-def setup(observational_data):
-    # 1. Read in the Causal DAG
-    causal_dag = CausalDAG(f"{ROOT}/dag.dot")
-
-    # 2. Create variables
-    pop_size = Input("pop_size", int)
-    pop_infected = Input("pop_infected", int)
-    n_days = Input("n_days", int)
-    cum_infections = Output("cum_infections", int)
-    cum_deaths = Output("cum_deaths", int)
-    location = Input("location", str)
-    variants = Input("variants", str)
-    avg_age = Input("avg_age", float)
-    beta = Input("beta", float)
-    contacts = Input("contacts", float)
-
-    # 3. Create scenario by applying constraints over a subset of the input variables
-    scenario = Scenario(
-        variables={
-            pop_size,
[…]
