### 2025-02-17 736cf6d
Removed data collector from metamorphic relation
