+++ tests/testing_tests/test_metamorphic_relations.py
@@ -1,0 +1,247 @@
+import unittest
+import os
+import shutil, tempfile
+import pandas as pd
+from itertools import combinations
+
+from causal_testing.specification.causal_dag import CausalDAG
+from causal_testing.specification.causal_specification import Scenario
+from causal_testing.testing.metamorphic_relation import (
+    ShouldCause,
+    ShouldNotCause,
+    generate_metamorphic_relations,
+    generate_metamorphic_relation,
+)
+from causal_testing.specification.variable import Input, Output
+from causal_testing.testing.base_test_case import BaseTestCase
+
+
+class TestMetamorphicRelation(unittest.TestCase):
+    def setUp(self) -> None:
+        self.temp_dir_path = tempfile.mkdtemp()
+        self.dag_dot_path = os.path.join(self.temp_dir_path, "dag.dot")
+        dag_dot = """digraph DAG { rankdir=LR; X1 -> Z; Z -> M; M -> Y; X2 -> Z; X3 -> M;}"""
+        with open(self.dag_dot_path, "w") as f:
+            f.write(dag_dot)
+        self.dcg_dot_path = os.path.join(self.temp_dir_path, "dcg.dot")
+        dcg_dot = """digraph dct { a -> b -> c -> d; d -> c; }"""
+        with open(self.dcg_dot_path, "w") as f:
+            f.write(dcg_dot)
+
+        X1 = Input("X1", float)
+        X2 = Input("X2", float)
+        X3 = Input("X3", float)
+        Z = Output("Z", float)
+        M = Output("M", float)
+        Y = Output("Y", float)
+        self.scenario = Scenario(variables={X1, X2, X3, Z, M, Y})
+        self.default_control_input_config = {"X1": 1, "X2": 2, "X3": 3}
+        self.default_treatment_input_config = {"X1": 2, "X2": 3, "X3": 3}
+
+    def tearDown(self) -> None:
+        shutil.rmtree(self.temp_dir_path)
+
+    def test_should_not_cause_json_stub(self):
+        """Test if the ShouldCause MR passes all metamorphic tests where the DAG perfectly represents the program
+        and there is only a single input."""
+        causal_dag = CausalDAG(self.dag_dot_path)
+        causal_dag.graph.remove_nodes_from(["X2", "X3"])
+        adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
+        should_not_cause_MR = ShouldNotCause(BaseTestCase("X1", "Z"), adj_set, causal_dag)
+        self.assertEqual(
+            should_not_cause_MR.to_json_stub(),
+            {
+                "effect": "direct",
+                "estimate_type": "coefficient",
+                "estimator": "LinearRegressionEstimator",
+                "expected_effect": {"Z": "NoEffect"},
+                "formula": "Z ~ X1",
+                "mutations": ["X1"],
+                "name": "X1 _||_ Z",
+                "formula": "Z ~ X1",
+                "alpha": 0.05,
+                "skip": True,
+            },
+        )
+
+    def test_should_cause_json_stub(self):
+        """Test if the ShouldCause MR passes all metamorphic tests where the DAG perfectly represents the program
+        and there is only a single input."""
+        causal_dag = CausalDAG(self.dag_dot_path)
+        causal_dag.graph.remove_nodes_from(["X2", "X3"])
+        adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
+        should_cause_MR = ShouldCause(BaseTestCase("X1", "Z"), adj_set, causal_dag)
[…]
