+++ causal_testing/testing/metamorphic_relation.py
@@ -13,11 +13,9 @@ import json
 from multiprocessing import Pool
 
 import networkx as nx
-import pandas as pd
-import numpy as np
 
 from causal_testing.specification.causal_specification import CausalDAG, Node
-from causal_testing.data_collection.data_collector import ExperimentalDataCollector
+from causal_testing.testing.base_test_case import BaseTestCase
 
 logger = logging.getLogger(__name__)
 
@@ -26,90 +24,10 @@ logger = logging.getLogger(__name__)
 class MetamorphicRelation:
     """Class representing a metamorphic relation."""
 
-    treatment_var: Node
-    output_var: Node
+    base_test_case: BaseTestCase
     adjustment_vars: Iterable[Node]
     dag: CausalDAG
     tests: Iterable = None
-
-    def generate_follow_up(self, n_tests: int, min_val: float, max_val: float, seed: int = 0):
-        """Generate numerical follow-up input configurations."""
-        np.random.seed(seed)
-
-        # Get set of variables to change, excluding the treatment itself
-        variables_to_change = {node for node in self.dag.graph.nodes if self.dag.graph.in_degree(node) == 0}
-        if self.adjustment_vars:
-            variables_to_change |= set(self.adjustment_vars)
-        if self.treatment_var in variables_to_change:
-            variables_to_change.remove(self.treatment_var)
-
-        # Assign random numerical values to the variables to change
-        test_inputs = pd.DataFrame(
-            np.random.randint(min_val, max_val, size=(n_tests, len(variables_to_change))),
-            columns=sorted(variables_to_change),
-        )
-
-        # Enumerate the possible source, follow-up pairs for the treatment
-        candidate_source_follow_up_pairs = np.array(list(combinations(range(int(min_val), int(max_val + 1)), 2)))
-
-        # Sample without replacement from the possible source, follow-up pairs
-        sampled_source_follow_up_indices = np.random.choice(
-            candidate_source_follow_up_pairs.shape[0], n_tests, replace=False
-        )
-
-        follow_up_input = f"{self.treatment_var}'"
-        source_follow_up_test_inputs = pd.DataFrame(
-            candidate_source_follow_up_pairs[sampled_source_follow_up_indices],
-            columns=sorted([self.treatment_var] + [follow_up_input]),
-        )
-        self.tests = [
-            MetamorphicTest(
-                source_inputs,
-                follow_up_inputs,
-                other_inputs,
-                self.output_var,
-                str(self),
-            )
-            for source_inputs, follow_up_inputs, other_inputs in zip(
-                source_follow_up_test_inputs[[self.treatment_var]].to_dict(orient="records"),
-                source_follow_up_test_inputs[[follow_up_input]]
-                .rename(columns={follow_up_input: self.treatment_var})
-                .to_dict(orient="records"),
-                (
-                    test_inputs.to_dict(orient="records")
-                    if not test_inputs.empty
-                    else [{}] * len(source_follow_up_test_inputs)
-                ),
-            )
-        ]
-
-    def execute_tests(self, data_collector: ExperimentalDataCollector):
-        """Execute the generated list of metamorphic tests, returning a dictionary of tests that pass and fail.
-
-        :param data_collector: An experimental data collector for the system-under-test.
-        """
-        test_results = {"pass": [], "fail": []}
-        for metamorphic_test in self.tests:
-            # Update the control and treatment configuration to take generated values for source and follow-up tests
[…]

[…]
