### 2025-02-17 73eb5f1
Removed unnecessary methods and arguments from metamorphic relation
