+++ tests/testing_tests/test_metamorphic_relations.py
@@ -47,7 +47,7 @@ class TestMetamorphicRelation(unittest.TestCase):
         causal_dag = CausalDAG(self.dag_dot_path)
         causal_dag.graph.remove_nodes_from(["X2", "X3"])
         adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
-        should_not_cause_MR = ShouldNotCause(BaseTestCase("X1", "Z"), adj_set, causal_dag)
+        should_not_cause_MR = ShouldNotCause(BaseTestCase("X1", "Z"), adj_set)
         self.assertEqual(
             should_not_cause_MR.to_json_stub(),
             {
@@ -70,7 +70,7 @@ class TestMetamorphicRelation(unittest.TestCase):
         causal_dag = CausalDAG(self.dag_dot_path)
         causal_dag.graph.remove_nodes_from(["X2", "X3"])
         adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
-        should_cause_MR = ShouldCause(BaseTestCase("X1", "Z"), adj_set, causal_dag)
+        should_cause_MR = ShouldCause(BaseTestCase("X1", "Z"), adj_set)
         self.assertEqual(
             should_cause_MR.to_json_stub(),
             {
@@ -94,12 +94,12 @@ class TestMetamorphicRelation(unittest.TestCase):
 
         # Check all ShouldCause relations are present and no extra
         expected_should_cause_relations = [
-            ShouldCause(BaseTestCase("X1", "Z"), [], dag),
-            ShouldCause(BaseTestCase("Z", "M"), [], dag),
-            ShouldCause(BaseTestCase("M", "Y"), ["Z"], dag),
-            ShouldCause(BaseTestCase("Z", "Y"), ["M"], dag),
-            ShouldCause(BaseTestCase("X2", "Z"), [], dag),
-            ShouldCause(BaseTestCase("X3", "M"), [], dag),
+            ShouldCause(BaseTestCase("X1", "Z"), []),
+            ShouldCause(BaseTestCase("Z", "M"), []),
+            ShouldCause(BaseTestCase("M", "Y"), ["Z"]),
+            ShouldCause(BaseTestCase("Z", "Y"), ["M"]),
+            ShouldCause(BaseTestCase("X2", "Z"), []),
+            ShouldCause(BaseTestCase("X3", "M"), []),
         ]
 
         extra_sc_relations = [scr for scr in should_cause_relations if scr not in expected_should_cause_relations]
@@ -110,15 +110,15 @@ class TestMetamorphicRelation(unittest.TestCase):
 
         # Check all ShouldNotCause relations are present and no extra
         expected_should_not_cause_relations = [
-            ShouldNotCause(BaseTestCase("X1", "X2"), [], dag),
-            ShouldNotCause(BaseTestCase("X1", "X3"), [], dag),
-            ShouldNotCause(BaseTestCase("X1", "M"), ["Z"], dag),
-            ShouldNotCause(BaseTestCase("X1", "Y"), ["Z"], dag),
-            ShouldNotCause(BaseTestCase("X2", "X3"), [], dag),
-            ShouldNotCause(BaseTestCase("X2", "M"), ["Z"], dag),
-            ShouldNotCause(BaseTestCase("X2", "Y"), ["Z"], dag),
-            ShouldNotCause(BaseTestCase("X3", "Y"), ["M", "Z"], dag),
-            ShouldNotCause(BaseTestCase("Z", "X3"), [], dag),
[…]
