### 2025-02-17 127a2f4
Added experimental estimator to keep functionality of experimental data collector
