+++ causal_testing/estimation/experimental_estimator.py
@@ -1,8 +1,8 @@
 """This module contains the ExperimentalEstimator class for directly interacting with the system under test."""
 
-import pandas as pd
 from typing import Any
 from abc import abstractmethod
+import pandas as pd
 
 from causal_testing.estimation.abstract_estimator import Estimator
 
