### 2025-02-17 7e33f80
fix: schematic diagram on homepage
