+++ docs/source/conf.py
@@ -50,7 +50,7 @@ html_theme = "sphinx_rtd_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ['_static', os.path.abspath('../../images')] # add /images directory to static path
 
 html_css_files = ['css/custom.css']
 
