### 2025-02-17 61d1e3d
fix: remove mentions of data collector
