### 2025-02-17 #7739 @yinfangchen
Sorry for the delay. I think this software satisfies most of the requirements.

I do encounter a problem when trying to run the test cases: https://causal-testing-framework.readthedocs.io/en/latest/frontends/test_suite.html#execution

It seems that the `execute_test_suite` is not under the `causal_test_engine`. I cannot find the `causal_test_engine` anywhere in the current repo. If there is a version mismatch, please update.
