+++ docs/source/usage.rst
@@ -107,4 +107,4 @@ various information. Here, we simply assert that the observed result is (on aver
    test_passes = causal_test_case.expected_causal_effect.apply(causal_test_result)
    assert test_passes, "Expected to see a positive change in y."
 
-Multiple tests can be executed at once using the test engines :doc:`Test Suite </frontends/test_suite>` feature.
+Multiple tests can be executed at once using a causal test suite :doc:`Test Suite </frontends/test_suite>` feature.
