### 2025-02-18 2786fdc
Fixed metamorphic relation
