### 2025-02-18 c38706d
Fixed default behaviour of MR generation
