+++ examples/poisson-line-process/example_pure_python.py
@@ -1,0 +1,216 @@
+import os
+import logging
+
+import pandas as pd
+
+from causal_testing.specification.causal_dag import CausalDAG
+from causal_testing.specification.scenario import Scenario
+from causal_testing.specification.variable import Input, Output
+from causal_testing.specification.causal_specification import CausalSpecification
+from causal_testing.testing.causal_test_case import CausalTestCase
+from causal_testing.testing.causal_test_outcome import ExactValue, Positive
+from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
+from causal_testing.estimation.abstract_estimator import Estimator
+from causal_testing.testing.base_test_case import BaseTestCase
+from causal_testing.testing.causal_test_suite import CausalTestSuite
+
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.DEBUG, format="%(message)s")
+
+
+class EmpiricalMeanEstimator(Estimator):
+    def add_modelling_assumptions(self):
+        """
+        Add modelling assumptions to the estimator. This is a list of strings which list the modelling assumptions that
+        must hold if the resulting causal inference is to be considered valid.
+        """
+        self.modelling_assumptions += "The data must contain runs with the exact configuration of interest."
+
+    def estimate_ate(self) -> float:
+        """Estimate the outcomes under control and treatment.
+        :return: The empirical average treatment effect.
+        """
+        control_results = self.df.where(self.df[self.treatment] == self.control_value)[self.outcome].dropna()
+        treatment_results = self.df.where(self.df[self.treatment] == self.treatment_value)[self.outcome].dropna()
+        return treatment_results.mean() - control_results.mean(), None
+
+    def estimate_risk_ratio(self) -> float:
+        """Estimate the outcomes under control and treatment.
+        :return: The empirical average treatment effect.
+        """
+        control_results = self.df.where(self.df[self.treatment] == self.control_value)[self.outcome].dropna()
+        treatment_results = self.df.where(self.df[self.treatment] == self.treatment_value)[self.outcome].dropna()
+        return treatment_results.mean() / control_results.mean(), None
+
+
+# 1. Read in the Causal DAG
+ROOT = os.path.realpath(os.path.dirname(__file__))
+causal_dag = CausalDAG(f"{ROOT}/dag.dot")
+
+# 2. Create variables
+width = Input("width", float)
+height = Input("height", float)
+intensity = Input("intensity", float)
+
+num_lines_abs = Output("num_lines_abs", float)
+num_lines_unit = Output("num_lines_unit", float)
+num_shapes_abs = Output("num_shapes_abs", float)
+num_shapes_unit = Output("num_shapes_unit", float)
+
+# 3. Create scenario
+scenario = Scenario(
+    variables={
+        width,
+        height,
+        intensity,
+        num_lines_abs,
+        num_lines_unit,
+        num_shapes_abs,
+        num_shapes_unit,
+    }
+)
+
+# 4. Construct a causal specification from the scenario and causal DAG
+causal_specification = CausalSpecification(scenario, causal_dag)
+
+observational_data_path = f"{ROOT}/data/random/data_random_1000.csv"
+
+
+def causal_test_intensity_num_shapes(
+    observational_data_path,
+    causal_test_case,
+    square_terms=[],
+    inverse_terms=[],
+    empirical=False,
+):
+    # 8. Set up an estimator
+    data = pd.read_csv(observational_data_path, index_col=0).astype(float)
+
+    treatment = causal_test_case.treatment_variable.name
+    outcome = causal_test_case.outcome_variable.name
+
+    estimator = None
+    if empirical:
+        estimator = EmpiricalMeanEstimator(
+            treatment=[treatment],
[…]
