### 2025-02-18 6071bd3
Reworked examples
