+++ examples/poisson/example_run_causal_tests.py
@@ -1,183 +1,0 @@
-import numpy as np
-import pandas as pd
-import scipy
-import os
-
-from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
-from causal_testing.testing.causal_test_outcome import ExactValue, Positive, Negative, NoEffect, CausalTestOutcome
-from causal_testing.testing.causal_test_result import CausalTestResult
-from causal_testing.json_front.json_class import JsonUtility
-from causal_testing.specification.scenario import Scenario
-from causal_testing.specification.variable import Input, Output, Meta
-
-import logging
-
-logger = logging.getLogger(__name__)
-logging.basicConfig(level=logging.DEBUG, format="%(message)s")
-
-
-class WidthHeightEstimator(LinearRegressionEstimator):
-    """
-    Extension of LinearRegressionEstimator class to include scenario specific user code
-    """
-
-    def add_modelling_assumptions(self):
-        """
-        Add modelling assumptions to the estimator. This is a list of strings which list the modelling assumptions that
-        must hold if the resulting causal inference is to be considered valid.
-        """
-        self.modelling_assumptions += "The output varies according to 2i(w+h)"
-
-    def estimate_ate(self) -> (float, [float, float], float):
-        """Estimate the conditional average treatment effect of the treatment on the outcome. That is, the change
-        in outcome caused by changing the treatment variable from the control value to the treatment value.
-        :return: The conditional average treatment effect and the 95% Wald confidence intervals.
-        """
-        assert (
-            self.effect_modifiers
-        ), f"Must have at least one effect modifier to compute CATE - {self.effect_modifiers}."
-        x = pd.DataFrame()
-        x[self.treatment[0]] = [self.treatment_values, self.control_values]
-        x["Intercept"] = 1
-        for k, v in self.effect_modifiers.items():
-            self.adjustment_set.add(k)
-            x[k] = v
-        if hasattr(self, "product_terms"):
-            for a, b in self.product_terms:
-                x[f"{a}*{b}"] = x[a] * x[b]
-
-        x.drop(["width", "height"], axis=1, inplace=True)
-        self.adjustment_set = {"width*intensity", "height*intensity"}
-
-        logger.info("%s", x)
-        logger.info("%s", self.adjustment_set)
-
-        model = self._run_linear_regression()
-        logger.info("%s", model.summary())
-        y = model.predict(x)
-        treatment_outcome = y.iloc[0]
-        control_outcome = y.iloc[1]
-
-        return treatment_outcome - control_outcome, None
-
-
-class PoissonWidthHeight(CausalTestOutcome):
-    """An extension of TestOutcome representing that the expected causal effect should be positive."""
-
-    def __init__(self, atol=0.5):
-        self.atol = atol
-        self.i2c = None
-
-    def apply(self, res: CausalTestResult) -> bool:
-        # TODO: confidence intervals?
-        logger.info("=== APPLYING ===")
-        logger.info("effect_modifier_configuration", res.effect_modifier_configuration)
-        effect_modifier_configuration = {k.name: v for k, v in res.effect_modifier_configuration.items()}
-        c = res.treatment_value - res.control_value
-        i = effect_modifier_configuration["intensity"]
-        self.i2c = i * 2 * c
-        logger.info("2ic: 2 * %s * %s = %s", i, c, self.i2c)
-        logger.info("ate: %s", res.test_value.value)
[…]
