+++ examples/poisson-line-process/example_poisson_process.py
@@ -1,210 +1,0 @@
-from causal_testing.specification.causal_dag import CausalDAG
-from causal_testing.specification.scenario import Scenario
-from causal_testing.specification.variable import Input, Output
-from causal_testing.specification.causal_specification import CausalSpecification
-from causal_testing.testing.causal_test_case import CausalTestCase
-from causal_testing.testing.causal_test_outcome import ExactValue, Positive
-from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
-from causal_testing.estimation.abstract_estimator import Estimator
-from causal_testing.testing.base_test_case import BaseTestCase
-
-import pandas as pd
-import os
-import logging
-
-logger = logging.getLogger(__name__)
-logging.basicConfig(level=logging.DEBUG, format="%(message)s")
-
-
-class EmpiricalMeanEstimator(Estimator):
-    def add_modelling_assumptions(self):
-        """
-        Add modelling assumptions to the estimator. This is a list of strings which list the modelling assumptions that
-        must hold if the resulting causal inference is to be considered valid.
-        """
-        self.modelling_assumptions += "The data must contain runs with the exact configuration of interest."
-
-    def estimate_ate(self) -> float:
-        """Estimate the outcomes under control and treatment.
-        :return: The empirical average treatment effect.
-        """
-        control_results = self.df.where(self.df[self.treatment[0]] == self.control_value)[self.outcome].dropna()
-        treatment_results = self.df.where(self.df[self.treatment[0]] == self.treatment_value)[self.outcome].dropna()
-        return treatment_results.mean()[0] - control_results.mean()[0], None
-
-    def estimate_risk_ratio(self) -> float:
-        """Estimate the outcomes under control and treatment.
-        :return: The empirical average treatment effect.
-        """
-        control_results = self.df.where(self.df[self.treatment[0]] == self.control_value)[self.outcome].dropna()
-        treatment_results = self.df.where(self.df[self.treatment[0]] == self.treatment_value)[self.outcome].dropna()
-        return treatment_results.mean()[0] / control_results.mean()[0], None
-
-
-# 1. Read in the Causal DAG
-ROOT = os.path.realpath(os.path.dirname(__file__))
-causal_dag = CausalDAG(f"{ROOT}/dag.dot")
-
-# 2. Create variables
-width = Input("width", float)
-height = Input("height", float)
-intensity = Input("intensity", float)
-
-num_lines_abs = Output("num_lines_abs", float)
-num_lines_unit = Output("num_lines_unit", float)
-num_shapes_abs = Output("num_shapes_abs", float)
-num_shapes_unit = Output("num_shapes_unit", float)
-
-# 3. Create scenario by applying constraints over a subset of the input variables
-scenario = Scenario(
-    variables={
-        width,
-        height,
-        intensity,
-        num_lines_abs,
-        num_lines_unit,
-        num_shapes_abs,
-        num_shapes_unit,
-    }
-)
-
-# 4. Construct a causal specification from the scenario and causal DAG
-causal_specification = CausalSpecification(scenario, causal_dag)
-
-observational_data_path = f"{ROOT}/data/random/data_random_1000.csv"
-
-
-def causal_test_intensity_num_shapes(
-    observational_data_path,
-    causal_test_case,
-    square_terms=[],
-    inverse_terms=[],
-    empirical=False,
-):
-    # 7. Obtain the minimal adjustment set for the causal test case from the causal DAG
-    minimal_adjustment_set = causal_dag.identification(causal_test_case.base_test_case)
-
-    # 8. Set up an estimator
-    data = pd.read_csv(observational_data_path, index_col=0).astype(float)
-
-    treatment = causal_test_case.treatment_variable.name
-    outcome = causal_test_case.outcome_variable.name
-
-    estimator = None
[…]
