+++ tests/json_front_tests/test_json_class.py
@@ -1,4 +1,5 @@
 import unittest
+import pytest
 from pathlib import Path
 from statistics import StatisticsError
 import scipy
@@ -13,6 +14,7 @@ from causal_testing.specification.scenario import Scenario
 from causal_testing.specification.causal_specification import CausalSpecification
 
 
+@pytest.mark.skip(reason="json frontend under reconstruction")
 class TestJsonClass(unittest.TestCase):
     """Test the JSON frontend for the Causal Testing Framework (CTF)
 
@@ -88,7 +90,7 @@ class TestJsonClass(unittest.TestCase):
                     "name": "test1",
                     "mutations": {"test_input": "Increase"},
                     "estimator": "LinearRegressionEstimator",
-                    "estimate_type": "ate",
+                    "estimate_type": "coefficient",
                     "effect_modifiers": [],
                     "expected_effect": {"test_output": "NoEffect"},
                     "skip": False,
@@ -165,7 +167,7 @@ class TestJsonClass(unittest.TestCase):
                     "name": "test1",
                     "mutations": {"test_input_no_dist": "Increase"},
                     "estimator": "LinearRegressionEstimator",
-                    "estimate_type": "ate",
+                    "estimate_type": "coefficient",
                     "effect_modifiers": [],
                     "expected_effect": {"test_output": "NoEffect"},
                     "skip": False,
@@ -194,7 +196,7 @@ class TestJsonClass(unittest.TestCase):
                     "name": "test1",
                     "mutations": {"test_input": "Increase"},
                     "estimator": "LinearRegressionEstimator",
-                    "estimate_type": "ate",
+                    "estimate_type": "coefficient",
                     "effect_modifiers": [],
                     "expected_effect": {"test_output": "Positive"},
                     "skip": False,
@@ -224,7 +226,7 @@ class TestJsonClass(unittest.TestCase):
                     "control_value": 0,
                     "treatment_value": 1,
                     "estimator": "LinearRegressionEstimator",
-                    "estimate_type": "ate",
+                    "estimate_type": "coefficient",
                     "expected_effect": {"test_output": "NoEffect"},
                     "skip": False,
                 }
@@ -239,38 +241,6 @@ class TestJsonClass(unittest.TestCase):
             temp_out = reader.readlines()
         self.assertIn("FAILED", temp_out[-1])
 
-    def test_concrete_generate_params(self):
-        example_test = {
-            "tests": [
-                {
-                    "name": "test1",
-                    "mutations": {"test_input": "Increase"},
-                    "estimator": "LinearRegressionEstimator",
-                    "estimate_type": "ate",
-                    "effect_modifiers": [],
-                    "expected_effect": {"test_output": "NoEffect"},
-                    "sample_size": 5,
-                    "target_ks_score": 0.05,
-                    "skip": False,
-                }
-            ]
-        }
-        self.json_class.test_plan = example_test
-        effects = {"NoEffect": NoEffect()}
-        mutates = {
-            "Increase": lambda x: self.json_class.scenario.treatment_variables[x].z3
-            > self.json_class.scenario.variables[x].z3
-        }
-        estimators = {"LinearRegressionEstimator": LinearRegressionEstimator}
-
-        self.json_class.run_json_tests(effects=effects, estimators=estimators, f_flag=False, mutates=mutates)
-
-        # Test that the final log message prints that failed tests are printed, which is expected behaviour for this
-        # scenario
-        with open("temp_out.txt", "r") as reader:
-            temp_out = reader.readlines()
-        self.assertIn("failed", temp_out[-1])
-
     def test_no_data_provided(self):
         example_test = {
             "tests": [
@@ -278,7 +248,7 @@ class TestJsonClass(unittest.TestCase):
                     "name": "test1",
[…]
