### 2025-02-18 7c0f0f2
Removed abstract causal test case + pylint
