+++ causal_testing/generation/abstract_causal_test_case.py
@@ -1,279 +1,0 @@
-"""This module contains the class AbstractCausalTestCase, which generates concrete test cases"""
-
-import itertools
-import logging
-from enum import Enum
-from typing import Iterable
-
-import lhsmdu
-import pandas as pd
-import z3
-from scipy import stats
-
-
-from causal_testing.specification.scenario import Scenario
-from causal_testing.specification.variable import Variable
-from causal_testing.testing.causal_test_case import CausalTestCase
-from causal_testing.testing.causal_test_outcome import CausalTestOutcome
-from causal_testing.testing.base_test_case import BaseTestCase
-
-
-logger = logging.getLogger(__name__)
-
-
-class AbstractCausalTestCase:
-    """
-    An abstract test case serves as a generator for concrete test cases. Instead of having concrete control
-    and treatment values, we instead just specify the intervention and the treatment variables. This then
-    enables potentially infinite concrete test cases to be generated between different values of the treatment.
-    """
-
-    def __init__(
-        # pylint: disable=too-many-arguments
-        self,
-        scenario: Scenario,
-        intervention_constraints: set[z3.ExprRef],
-        treatment_variable: Variable,
-        expected_causal_effect: dict[Variable:CausalTestOutcome],
-        effect_modifiers: set[Variable] = None,
-        estimate_type: str = "ate",
-        effect: str = "total",
-    ):
-        if treatment_variable not in scenario.variables.values():
-            raise ValueError(
-                "Treatment variables must be a subset of variables."
-                + f" Instead got:\ntreatment_variables={treatment_variable}\nvariables={scenario.variables}"
-            )
-
-        assert len(expected_causal_effect) == 1, "We currently only support tests with one causal outcome"
-
-        self.scenario = scenario
-        self.intervention_constraints = intervention_constraints
-        self.treatment_variable = treatment_variable
-        self.expected_causal_effect = expected_causal_effect
-        self.estimate_type = estimate_type
-        self.effect = effect
-
-        if effect_modifiers is not None:
-            self.effect_modifiers = effect_modifiers
-        else:
-            self.effect_modifiers = {}
-
-    def __str__(self):
-        outcome_string = " and ".join(
-            [f"the effect on {var} should be {str(effect)}" for var, effect in self.expected_causal_effect.items()]
-        )
-        return f"When we apply intervention {self.intervention_constraints}, {outcome_string}"
-
-    def datapath(self) -> str:
-        """Create and return the sanitised data path"""
-
-        def sanitise(string):
-            return "".join([x for x in string if x.isalnum()])
-
-        return (
-            sanitise("-".join([str(c) for c in self.intervention_constraints]))
-            + "_"
-            + "-".join([f"{v.name}_{e}" for v, e in self.expected_causal_effect.items()])
-            + ".csv"
-        )
-
-    def _generate_concrete_tests(
-        # pylint: disable=too-many-locals
-        self,
-        sample_size: int,
-        rct: bool = False,
-        seed: int = 0,
-    ) -> tuple[list[CausalTestCase], pd.DataFrame]:
-        """Generates a list of `num` concrete test cases.
-
-        :param sample_size: The number of strata to use for Latin hypercube sampling. Where no target_ks_score is
-        provided, this corresponds to the number of test cases to generate. Where target_ks_score is provided, the
-        number of test cases will be a multiple of this.
-        :param rct: Whether we're running an RCT, i.e. whether to add the treatment run to the concrete runs.
-        :param seed: Random seed for reproducability.
-        :return: A list of causal test cases and a dataframe representing the required model run configurations.
-        :rtype: ([CausalTestCase], pd.DataFrame)
-        """
-
-        concrete_tests = []
-        runs = []
-        run_columns = sorted([v.name for v in self.scenario.variables.values() if v.distribution])
-
[…]
