+++ tests/specification_tests/test_variable.py
@@ -1,9 +1,8 @@
 import unittest
 from enum import Enum
-import z3
 from scipy.stats import norm, kstest
 
-from causal_testing.specification.variable import z3_types, Variable, Input
+from causal_testing.specification.variable import Variable, Input
 
 
 class TestVariable(unittest.TestCase):
@@ -14,116 +13,9 @@ class TestVariable(unittest.TestCase):
     def setUp(self) -> None:
         pass
 
-    def test_z3_types_enum(self):
-        class Color(Enum):
-            """
-            Example enum class color.
-            """
-
-            RED = "RED"
-            GREEN = "GREEN"
-            BLUE = "BLUE"
-
-        dtype, _ = z3.EnumSort("color", ("RED", "GREEN", "BLUE"))
-        z3_color = z3.Const("color", dtype)
-        expected = z3_types(Color)("color")
-        # No actual way to assert their equality since they are two different objects
-        expected_values = [expected.sort().constructor(c)() for c in range(expected.sort().num_constructors())]
-        z3_color_values = [z3_color.sort().constructor(c)() for c in range(z3_color.sort().num_constructors())]
-
-        # This isn't very good, but I think it's the best we can do since even
-        # z3_types(Color)("color") != z3_types(Color)("color")
-        self.assertEqual(list(map(str, expected_values)), list(map(str, z3_color_values)))
-
-    def test_cast_z3_bool(self):
-        bip = Input("bip", bool)
-        s = z3.Solver()
-        t = z3.Bool("t")
-        f = z3.Bool("f")
-        s.add(t)
-        s.add(z3.Not(f))
-        s.check()
-        self.assertEqual(bip.cast(s.model()[t]), True)
-        self.assertEqual(bip.cast(s.model()[f]), False)
-
-    def test_cast_z3_string(self):
-        ip = Input("bip", str)
-        s = z3.Solver()
-        t = z3.String("t")
-        s.add(t == "hello")
-        s.check()
-        self.assertEqual(ip.cast(s.model()[t]), "hello")
-
     def test_sample_flakey(self):
         ip = Input("ip", float, norm)
         self.assertGreater(kstest(ip.sample(10), norm.cdf).pvalue, 0.95)
-
-    def test_cast_enum(self):
-        class Color(Enum):
-            """
-            Example enum class color.
-            """
-
-            RED = "RED"
-            GREEN = "GREEN"
-            BLUE = "BLUE"
-
-        color = Input("color", Color)
-
-        dtype, colours = z3.EnumSort("color", ("RED", "GREEN", "BLUE"))
-        self.assertEqual(color.cast(colours[0]), Color.RED)
-
-    def test_z3_value_enum(self):
-        class Color(Enum):
-            """
-            Example enum class color.
-            """
-
-            RED = "RED"
-            GREEN = "GREEN"
-            BLUE = "BLUE"
-
-        dtype, members = z3.EnumSort("color", ("RED", "GREEN", "BLUE"))
-        z3_color = z3.Const("color", dtype)
-        color = Input("color", Color)
-
-        self.assertEqual(color.z3_val(z3_color, "RED"), members[0])
-
-    def test_z3_types_custom(self):
-        class Color:
-            """
-            Example enum class color.
-            """
-
-            RED = 1
-            GREEN = 2
-            BLUE = 3
-
-            @classmethod
-            def to_z3(cls):
-                dtype, _ = z3.EnumSort("Color", ("RED", "GREEN", "BLUE"))
-                return lambda x: z3.Const(x, dtype)
-
-        dtype, _ = z3.EnumSort("color", ("RED", "GREEN", "BLUE"))
[…]
