### 2025-02-20 b86c584
All the tests pass and got rid of JSON front
