### 2025-02-20 4d48785
IPCW outcome is now an output
