+++ causal_testing/estimation/linear_regression_estimator.py
@@ -36,6 +36,7 @@ class LinearRegressionEstimator(RegressionEstimator):
         query: str = "",
     ):
         # pylint: disable=too-many-arguments
+        # pylint: disable=R0801
         super().__init__(
             base_test_case=base_test_case,
             treatment_value=treatment_value,
