### 2025-02-20 949c17d
Removed the docs for the deprecated frontends
