+++ causal_testing/main.py
@@ -1,0 +1,467 @@
+"""This module contains the classes that executes the Causal Testing Framework."""
+
+import argparse
+import json
+import logging
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Dict, Any, Optional, List, Union, Sequence
+
+import pandas as pd
+
+from .specification.causal_dag import CausalDAG
+from .specification.scenario import Scenario
+from .specification.variable import Input, Output
+from .specification.causal_specification import CausalSpecification
+from .testing.causal_test_case import CausalTestCase
+from .testing.base_test_case import BaseTestCase
+from .testing.causal_test_outcome import NoEffect, SomeEffect, Positive, Negative
+from .testing.causal_test_result import CausalTestResult
+from .estimation.linear_regression_estimator import LinearRegressionEstimator
+from .estimation.logistic_regression_estimator import LogisticRegressionEstimator
+
+logger = logging.getLogger(__name__)
+
+
+@dataclass
+class CausalTestingPaths:
+    """
+    Class for managing paths for causal testing inputs and outputs.
+
+    :param dag_path: Path to the DAG definition file
+    :param data_paths: List of paths to input data files
+    :param test_config_path: Path to the test configuration file
+    :param output_path: Path where test results will be written
+    """
+
+    dag_path: Path
+    data_paths: List[Path]
+    test_config_path: Path
+    output_path: Path
+
+    def __init__(
+        self,
+        dag_path: Union[str, Path],
+        data_paths: List[Union[str, Path]],
+        test_config_path: Union[str, Path],
+        output_path: Union[str, Path],
+    ):
+        self.dag_path = Path(dag_path)
+        self.data_paths = [Path(p) for p in data_paths]
+        self.test_config_path = Path(test_config_path)
+        self.output_path = Path(output_path)
+
+    def validate_paths(self) -> None:
+        """
+        Validate existence of all input paths and writability of output path.
+
+        :raises: FileNotFoundError if any required input file is missing.
+        """
+
+        if not self.dag_path.exists():
+            raise FileNotFoundError(f"DAG file not found: {self.dag_path}")
+
+        for data_path in self.data_paths:
+            if not data_path.exists():
+                raise FileNotFoundError(f"Data file not found: {data_path}")
+
+        if not self.test_config_path.exists():
+            raise FileNotFoundError(f"Test configuration file not found: {self.test_config_path}")
+
+        if not self.output_path.parent.exists():
+            self.output_path.parent.mkdir(parents=True)
+
+
+class CausalTestingFramework:
+    # pylint: disable=too-many-instance-attributes
+    """
+    Main class for running causal tests.
+
+    :param paths: CausalTestingPaths object containing required file paths
+    :param ignore_cycles: Flag to ignore cycles in the DAG
+    :param query: Optional query string to filter the input dataframe
+
+    """
+
+    def __init__(self, paths: CausalTestingPaths, ignore_cycles: bool = False, query: Optional[str] = None):
+        self.paths = paths
+        self.ignore_cycles = ignore_cycles
+        self.query = query
+
+        # These will be populated during setup
+        self.dag: Optional[CausalDAG] = None
+        self.data: Optional[pd.DataFrame] = None
+        self.variables: Dict[str, Any] = {"inputs": {}, "outputs": {}, "metas": {}}
+        self.scenario: Optional[Scenario] = None
+        self.causal_specification: Optional[CausalSpecification] = None
+        self.test_cases: Optional[List[CausalTestCase]] = None
+
+    def setup(self) -> None:
+        """
+        Set up the framework by loading DAG, runtime csv data, creating the scenario and causal specification.
+
+        :raises: FileNotFoundError if required files are missing
[…]
