### 2025-02-26 #7739 @akothen
Everything looks good to me. But I want to mention one thing that the authors may want to consider: it is probably better not to assume that people will read the paper and the documentation to understand and use the tool. So both of them have to be pretty self-contained. The paper looks good to me, and so does the documentation, but the documentation makes references to causal DAGs without depicting what they look like, etc (https://causal-testing-framework.readthedocs.io/en/latest/modules/causal_specification.html#causal-dag). I believe the point of the documentation is to have much more technical information that the general audience can understand and enables people to use and play around with the tool.
