### 2025-02-26 #7739 @danielskatz
Thanks @akothen - this makes sense as a suggestion to the authors. You might want to open an issue in their repo for this, as this review issue will close soon, I hope, with the article's publication
