### 2025-02-26 6d5290b
minor change in paper.md
