### 2025-02-27 b462aa1
Removed all mention of data collection and json front end
