### 2025-02-27 e676dc2
Some frontend improvements
