### 2025-03-04 06e199d
Poisson tests using frontend
