+++ causal_testing/main.py
@@ -103,7 +103,6 @@ class CausalTestingFramework:
         Set up the framework by loading DAG, runtime csv data, creating the scenario and causal specification.
 
         :raises: FileNotFoundError if required files are missing
-        :raises: Exception if setup process fails
         """
 
         logger.info("Setting up Causal Testing Framework...")
@@ -128,17 +127,11 @@ class CausalTestingFramework:
     def load_dag(self) -> CausalDAG:
         """
         Load the causal DAG from the specified file path.
-
-         :raises: Exception if DAG loading fails
         """
         logger.info(f"Loading DAG from {self.paths.dag_path}")
-        try:
-            dag = CausalDAG(str(self.paths.dag_path), ignore_cycles=self.ignore_cycles)
-            logger.info(f"DAG loaded with {len(dag.graph.nodes)} nodes and {len(dag.graph.edges)} edges")
-            return dag
-        except Exception as e:
-            logger.error(f"Failed to load DAG: {str(e)}")
-            raise
+        dag = CausalDAG(str(self.paths.dag_path), ignore_cycles=self.ignore_cycles)
+        logger.info(f"DAG loaded with {len(dag.graph.nodes)} nodes and {len(dag.graph.edges)} edges")
+        return dag
 
     def _read_dataframe(self, data_path):
         if str(data_path).endswith(".csv"):
@@ -152,33 +145,22 @@ class CausalTestingFramework:
 
         :param query: Optional pandas query string to filter the loaded data
         :return: Combined pandas DataFrame containing all loaded and filtered data
-        :raises: Exception if data loading or query application fails
         """
         logger.info(f"Loading data from {len(self.paths.data_paths)} source(s)")
 
-        try:
-            dfs = [self._read_dataframe(data_path) for data_path in self.paths.data_paths]
-            data = pd.concat(dfs, axis=0, ignore_index=True)
-            logger.info(f"Initial data shape: {data.shape}")
+        dfs = [self._read_dataframe(data_path) for data_path in self.paths.data_paths]
+        data = pd.concat(dfs, axis=0, ignore_index=True)
+        logger.info(f"Initial data shape: {data.shape}")
 
-            if query:
-                try:
-                    logger.info(f"Attempting to apply query: '{query}'")
-                    data = data.query(query)
-                except Exception as e:
-                    logger.error(f"Failed to apply query '{query}': {str(e)}")
-                    raise
+        if query:
+            logger.info(f"Attempting to apply query: '{query}'")
+            data = data.query(query)
 
-            return data
-        except Exception as e:
-            logger.error(f"Failed to load data: {str(e)}")
-            raise
+        return data
 
     def create_variables(self) -> None:
         """
         Create variable objects from DAG nodes based on their connectivity.
-
-
         """
         for node_name, node_data in self.dag.graph.nodes(data=True):
             if node_name not in self.data.columns and not node_data.get("hidden", False):
@@ -195,11 +177,7 @@ class CausalTestingFramework:
                 self.variables["outputs"][node_name] = Output(name=node_name, datatype=dtype)
 
     def create_scenario_and_specification(self) -> None:
-        """Create scenario and causal specification objects from loaded data.
-
-
-        :raises: ValueError if scenario constraints filter out all data points
-        """
+        """Create scenario and causal specification objects from loaded data."""
         # Create scenario
         all_variables = list(self.variables["inputs"].values()) + list(self.variables["outputs"].values())
         self.scenario = Scenario(variables=all_variables)
@@ -207,51 +185,17 @@ class CausalTestingFramework:
         # Set up treatment variables
         self.scenario.setup_treatment_variables()
 
-        # Apply scenario constraints to data
-        self.apply_scenario_constraints()
[…]
