### 2025-03-04 b37cf09
Added a couple more tests
