### 2025-03-04 4b531e1
feat: ability to run tests in batches
