### 2025-03-06 #7739 @jmafoster1
Hi @danielskatz,
Sorry for the delay on this. It took a while to get everything up on our institutional repository, but it is there now. The DOI is https://doi.org/10.15131/shef.data.28539398.v1. The tagged release is [v10.0.1](https://github.com/CITCOM-project/CausalTestingFramework/releases/tag/v10.0.1). Thanks for organising this.
