### 2025-03-06 9c7e4cd
Records the formula actually used in estimation reguardless of whether one was supplied
