### 2025-03-06 #7739 @jmafoster1
> 👋 [@jmafoster1](https://github.com/jmafoster1) - is this part also done now?
>
> > ... We'll address the documentation comment from [@akothen](https://github.com/akothen) ... Hopefully we can get that done this week.

@danielskatz thanks for the reminder! This is addressed in [this PR](https://github.com/CITCOM-project/CausalTestingFramework/pull/314), although we plan to do a complete overhaul and update of our docs and tutorials in the near future to include Jupyter notebook tutorials.
