+++ paper/paper.md
@@ -34,7 +34,7 @@ authors:
     orcid: 0000-0003-2134-6548
     affiliation: 1
   - name: Robert M. Hierons
-    orcid: 0000-0003-2134-6548
+    orcid: 0000-0002-4771-1446
     affiliation: 1
 affiliations:
  - name: University of Sheffield, UK
