### 2025-03-06 #7739 @jmafoster1
Also @danielskatz I also just noticed a typo in one of the author's DOIs in the metadata. This was my fault in the original submission, and I've just fixed it as part of the above PR. Sorry about that!
