### 2025-03-06 #7739 @danielskatz
Ok - let me reaccept this. I'm also unsure why the DOI doesn't resolve, but perhaps the reaccept will also fix that
