### 2025-03-06 5210cf3
feat: functionality for test-specific queries
