### 2025-03-06 e6f27fe
fix: allow global and test-specific queries
