### 2025-03-06 71b5c5f
fix: tests for global and test-specific queries
