### 2025-03-06 #7739 @danielskatz
@jmafoster1 - the DOI didn't get registered because two authors are listed with the same ORCID.  Please either fix this or remove the orcid line that is incorrect. We don't require ORCIDs for all authors, so they should either be correct or not listed

>  - name: Christopher Wild
    orcid: 0009-0009-1195-1497
    affiliation: 1
>  - name: Richard Somers
    orcid: 0009-0009-1195-1497
    affiliation: 1
